+++ src/youtube_automation/entrypoints.py
@@ -97,6 +97,7 @@ yt_workspace_guard = _make_entrypoint("youtube_automation.commands.channel.works
 yt_workflow_state = _make_entrypoint("youtube_automation.commands.collections.workflow_state_cli")
 yt_workflow_status = _make_entrypoint("youtube_automation.commands.collections.workflow_status")
 yt_hybrid_runner = _make_entrypoint("youtube_automation.commands.system.hybrid_runner")
+yt_media_acceptance = _make_entrypoint("youtube_automation.commands.media.media_acceptance")
 yt_channel_trend = _make_entrypoint("youtube_automation.commands.analytics.channel_trend")
 yt_collection_preflight = _make_entrypoint("youtube_automation.commands.collections.collection_preflight")
 yt_collection_serve = _make_entrypoint("youtube_automation.commands.collections.collection_serve")